pyproject.toml
[project]
name = "x8lucas8x-blog"
version = "0.1.0"
description = "Lucas Lira Gomes' personal blog."
readme = "README.md"
authors = [
  {name = "Lucas Lira Gomes", email = "contact@x8lucas8x.com"}
]
maintainers = [
  {name = "Lucas Lira Gomes", email = "contact@x8lucas8x.com"}
]
keywords = ["blog"]
requires-python = ">=3.13"
dependencies = [
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "markdown>=3.7",
    "markupsafe>=3.0.2",
    "minify-html>=0.16.4",
    "pillow>=11.1.0",
    "pyyaml>=6.0.2",
    "starlette>=0.46.1",
    "typer>=0.15.2",
    "uvicorn[standard]>=0.34.0",
    "wand>=0.6.13",
]

[project.urls]
Homepage = "https://x8lucas8x.com"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "scalene>=1.5.51",
]

[project.scripts]
x8lucas8x-blog = "src.main"